pyproject.toml
[project]
name = "mcp-getting-started"
version = "0.1.0"
description = "a mcp example package"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]",
    "openai",
    "gradio",
    "autogen-agentchat",
    "autogen-ext[openai]",
    "autogen-ext[mcp]",
    "autogen-ext[gradio]",
    "gradio",
    "python-dotenv"
]
